showEverything

+++ Kouvee/Models/Produk.cs
@@ -8,7 +8,7 @@ namespace Kouvee.Models
 {
     public class Produk
     {
-        public Produk(int iD_Produk, string nama_Produk, int stock, int min_Stock, float harga_BeliProduk, float harga_JualProduk, string gambar_Produk, int iD_Pegawai)
+        public Produk(int iD_Produk, string nama_Produk, int stock, int min_Stock, int harga_BeliProduk, int harga_JualProduk, string gambar_Produk, int iD_Pegawai)
         {
             ID_Produk = iD_Produk;
             Nama_Produk = nama_Produk;
@@ -24,8 +24,8 @@ namespace Kouvee.Models
         public string Nama_Produk { get; set; }
         public int Stock { get; set; }
         public int Min_Stock { get; set; }
-        public float Harga_BeliProduk { get; set; }
-        public float Harga_JualProduk { get; set; }
+        public int Harga_BeliProduk { get; set; }
+        public int Harga_JualProduk { get; set; }
         public string Gambar_Produk { get; set; }
         public int ID_Pegawai { get; set; }
     }
